Create I am currently learning Data and Business Analytics at AI Lead Tech online
After doing my Master's Degree in Animal Science and working as an Operations Director for many years, I felt the need to empower my analytical capabilities in order to make more informed decisions. Hence, I enrolled on this programme. This is a five-month course which covers all the main aspects of what one needs to know and understand to be able to function as a Data Analyst. By the end of the course, I will be able to extract data for any organization, from several departments, even in different format using My SQL, clean and transform the data in Power Query, Build Data Model, Build Measures (DAX), Analyse (Xcel) and Visualize (Power BI), We will also learn Python Programming Language and do a Capstone Project (Intenship). It is an exciting Journey and worth the investment.